devfreq: Introduce devfreq boost driver
This driver boosts enumerated devfreq devices upon input, and allows for
boosting specific devfreq devices on other custom events. The boost
frequencies for this driver should be set so that frame drops are near-zero
at the boosted frequencies and power consumption is minimized at said
frequencies. The goal of this driver is to provide an interface to achieve
optimal device performance by requesting boosts on key events, such as when
a frame is ready to rendered to the display.

Currently, support is only present for boosting the cpubw devfreq device,
but the driver is structured in a way that makes it easy to add support for
new boostable devfreq devices in the future.

Signed-off-by: Sultan Alsawaf <sultanxda@gmail.com>
Signed-off-by: kdrag0n <dragon@khronodragon.com>
(cherry picked from commit cd117e3eec585a04b7213ebc02c3fda484a75a90)
Signed-off-by: Enzaik <imernestodiaz@gmail.com>
(cherry picked from commit fb32754da0fc9d5caa9fcded1415a472ee4d7247)
Signed-off-by: enzaik <imernestodiaz@gmail.com>